requirements.txt
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.3
flask==3.0.0
flask-cors==4.0.0
lxml==5.1.0
openpyxl==3.1.2
dashscope>=1.14.0
python-dotenv>=1.0.0
